+++ commands/certificates.go
@@ -3,60 +3,52 @@ package commands
 import (
 	"bonvoy/consul"
 	"bonvoy/envoy"
-	"flag"
 	"fmt"
+	"github.com/spf13/cobra"
 	"strings"
 )
 
-type ExpiredCertificatesCommand struct {
-	fs *flag.FlagSet
-	consul consul.Client
-	name string
+type ExpiredCertificatesController struct {
+	ServiceName string
+	Consul consul.Client
 }
 
-func BuildExpiredCertificatesCommand() *ExpiredCertificatesCommand {
-	gc := &ExpiredCertificatesCommand{
-		fs: flag.NewFlagSet("certs-expired", flag.ContinueOnError),
-		consul: consul.NewClient(),
-	}
-	gc.fs.Arg(0)
-	return gc
+func BuildExpiredCertificatesController(rootCmd *cobra.Command) {
+	rootCmd.AddCommand(&cobra.Command{
+		Use: "certificates expired",
+		Short: "Show all expired certificates",
+		Long:  `Display all expired sidecar certificates as compared to the local Consul agent`,
+		Args: cobra.MinimumNArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			controller := ExpiredCertificatesController{
+				ServiceName: args[1],
+				Consul: consul.NewClient(),
+			}
+			return controller.Run()
+		},
+	})
 }
 
-func (g *ExpiredCertificatesCommand) Name() string {
-	return g.fs.Name()
-}
-
-func (g *ExpiredCertificatesCommand) Init(args []string) error {
-	return g.fs.Parse(args)
-}
-
-func (g *ExpiredCertificatesCommand) Run() error {
-	var name = g.name
-	if name == "" {
-		name = g.fs.Arg(0)
-	}
-
-	e, err := envoy.NewFromServiceName(name)
+func (c *ExpiredCertificatesController) Run() error {
+	e, err := envoy.NewFromServiceName(c.ServiceName)
 	if err != nil {
 		return err
 	}
 	data := e.Certificates().Get()
 
 	for _, certs := range data.Certificates {
-		for _, c := range certs.CertificateChain {
-			a := strings.Split(c.SubjectAltNames[0].Uri, "/")
+		for _, cert := range certs.CertificateChain {
+			a := strings.Split(cert.SubjectAltNames[0].Uri, "/")
 			svc := a[len(a)-1]
-			leaf := g.consul.Agent().GetConnectLeafCaCertificate(svc)
+			leaf := c.Consul.Agent().GetConnectLeafCaCertificate(svc)
 
-			if c.ExpirationTime != leaf.ValidBefore {
+			if cert.ExpirationTime != leaf.ValidBefore {
 				fmt.Println(svc)
 				fmt.Println("  Envoy Process ID:", e.Pid)
-				fmt.Printf("  Envoy Certificate Expiry: %s (%s days)\n", c.ExpirationTime, c.DaysUntilExpiration)
+				fmt.Printf("  Envoy Certificate Expiry: %s (%s days)\n", cert.ExpirationTime, cert.DaysUntilExpiration)
 				fmt.Println("  Consul Agent Certificate Expiry: ", leaf.ValidBefore)
 			}
 		}
 	}
-
 	return nil
 }