+++ commands/listeners.go
@@ -4,47 +4,42 @@ import (
 	"bonvoy/envoy"
 	"flag"
 	"fmt"
+	"github.com/spf13/cobra"
 )
 
-type ListenersCommand struct {
-	fs *flag.FlagSet
-	name string
+type ListenersController struct {
+	ServiceName string
 }
 
-// ListenersCommand
-func BuildListeners() *ListenersCommand {
-	gc := &ListenersCommand{
-		fs: flag.NewFlagSet("listeners", flag.ContinueOnError),
-	}
-	gc.fs.Arg(0)
-	gc.fs.StringVar(&gc.name, "service", "", "name of the service whose sidecar to enter")
-	return gc
+func BuildListenersCommand(rootCmd *cobra.Command) {
+	rootCmd.AddCommand(&cobra.Command{
+		Use: "listeners",
+		Short: "Show Envoy listeners",
+		Long:  `Display all registered Envoy sidecar listeners`,
+		Args: cobra.MinimumNArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			controller := ListenersController{
+				ServiceName: args[1],
+			}
+			return controller.Run()
+		},
+	})
 }
 
-func (g *ListenersCommand) Name() string {
-	return g.fs.Name()
-}
-
-func (g *ListenersCommand) Init(args []string) error {
-	return g.fs.Parse(args)
-}
-
-func (g *ListenersCommand) Run() error {
-	var name = g.name
-	if name == "" {
-		name = g.fs.Arg(0)
-	}
-
-	e, err := envoy.NewFromServiceName(name)
+func (s *ListenersController) Run() error {
+	e, err := envoy.NewFromServiceName(s.ServiceName)
 	if err != nil {
 		return err
 	}
 	listeners := e.Listeners().Get()
-
 	fmt.Println("LISTENERS:")
 	fmt.Println("----------------------------------------------------------------------")
 	for _, listener := range listeners {
 		fmt.Printf("%s: %s\n", listener.Name, listener.TargetAddress)
 	}
 	return nil
+}
+type ListenersCommand struct {
+	fs *flag.FlagSet
+	name string
 }