+++ CHANGELOG.md
@@ -2,8 +2,9 @@ The changelog for Bonvoy
 
 ## Pending Release
 
-* Add `server-info` command to display information about the Envoy sidecar
-* Add `log-level` command to set Envoy's log level
+* Restructured to use cobra cli library rather than flag
+* Add `server info` command to display information about the Envoy sidecar
+* Add `log level` command to set Envoy's log level
 * Restructure for better memory usage and a more closed API
 
 ## 0.0.1